+++ CHANGELOG.md
@@ -3,5 +3,13 @@
 - adding @all mention.
 - fixing history.
 - efecto gotitas de lluvia.
-- gecoding ip
-- whoisdata
+- gecoding ip in whois
+- fixing emotes: cascada
+- adding emotes: amor
+- Emotes renames: 
+    * miranda2 -> stalker
+    * sparkle -> hechizo
+    * stick -> varita
+    * wizard -> mago
+    * handshake -> amigos
+- Message split at 510.